requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.21
requests==2.31.0
python-dotenv==1.0.0
langchain-google-genai==0.0.1
pymysql==1.0.3
gunicorn
google-generativeai
langchain==0.1.13
langchain-google-genai==0.0.8
pydantic==2.6.4